+++ assets/templates/languages/nuxt/nuxt.config.ts
@@ -2,4 +2,10 @@
 export default defineNuxtConfig({
   compatibilityDate: "2026-01-01",
   devtools: { enabled: true },
+  modules: ["@nuxt/eslint"],
+  eslint: {
+    config: {
+      stylistic: true,
+    },
+  },
 });